Leave only `onError` callback to handle transport exceptions from WebSocket library
OnError handler in nv-websocket library is a base callback for all
errors (e.g. `onFrameError`, `onSendError` etc). We don't need to
override specific callbacks because the exception is passed to
`onError` callback first.

`onConnectError` is used only when connect method is asynchronous.
In SDK we are using synchoronios connect so `onConnectError` can't
be triggered at all.